Add with_border method to widget::TextBox
This method can be used to turn off border painting for the text box.
Useful if you want to customise the look of the text box.

The same thing can be achieved by setting the env's theme to colours
with full transparency, but doing so would involve unnecessary drawing
operations.